+++ WIP/Source/SourceCode/QLHS/BUL/ScoreBySubject_BUL.cs
@@ -25,7 +25,7 @@ namespace BUL
         {
             return ScoreBySubject_DAL.GetID(ProcessID, SubjectID);
         }
-        public static bool UpdateScore(string ID, float score)
+        public static bool UpdateScore(string ID, double score)
         {
             return ScoreBySubject_DAL.UpdateScore(ID, score);
         }
